% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_col_exists}
\alias{check_col_exists}
\title{Check That Required Columns Exist}
\usage{
check_col_exists(data, cols)
}
\arguments{
\item{data}{A data frame.}

\item{cols}{Character vector. Column names to check for.}
}
\value{
Invisible NULL. Throws an error if any column is missing.
}
\description{
Check That Required Columns Exist
}
\keyword{internal}
