% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_trait_cols}
\alias{detect_trait_cols}
\title{Auto-detect Numeric Trait Columns}
\usage{
detect_trait_cols(data, exclude = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{exclude}{Character vector. Additional column names to exclude.}
}
\value{
Character vector of numeric column names.
}
\description{
Returns the names of all numeric columns in a data frame, excluding
common non-trait columns like row, col, rep, block, and ID columns.
}
\examples{
\dontrun{
data(example_trial)
detect_trait_cols(example_trial)
# [1] "yield" "plant_height" "days_to_flower"
}

}
