% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoqc.R
\name{print.phenoqc_result}
\alias{print.phenoqc_result}
\title{Print a phenoQC Result}
\usage{
\method{print}{phenoqc_result}(x, ...)
}
\arguments{
\item{x}{A \code{phenoqc_result} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a phenoQC Result
}
